pyproject.toml
[project]
dynamic = ["version"]

dependencies = [
  "httpx>=0.28.1",
  "python-dotenv>=1.0.1",
  "requests>=2.32.3",
]
description = "Add your description here"
name = "withings-api"
readme = "README.md"
requires-python = ">=3.11"
[tool.setuptools.dynamic]
version = {attr="withings_api.__version__"}
[project.optional-dependencies]
all = ["withings-api[dev]"]
dev = [
  "openapi-python-client>=0.23.1",
  "ruff>=0.9.6",
  "tree-sitter>=0.24.0",
  "tree-sitter-javascript>=0.23.1",
  "types-pyyaml>=6.0.12.20241230",
  "types-requests>=2.32.0.20241016",
]

[dependency-groups]
dev = [
  "pyyaml>=6.0.2",
]

[tool.setuptools.package-dir]
withings_api = "withings_api"
